Refactored command classes to follow naming conventions.
