+++ src/Command/Repository/AbstractRepositoryCommand.php
@@ -26,10 +26,10 @@ use Baleen\Migrations\Version\Collection\LinkedVersions;
 use League\Flysystem\Filesystem;
 
 /**
- * Class RepositoryCommand
+ * Class AbstractRepositoryCommand
  * @author Gabriel Somoza <gabriel@strategery.io>
  */
-abstract class RepositoryCommand extends AbstractCommand
+abstract class AbstractRepositoryCommand extends AbstractCommand
 {
     /** @var RepositoryInterface */
     protected $repository;
