+++ src/Command/Repository/LatestCommand.php
@@ -27,7 +27,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  * Class ListCommand
  * @author Gabriel Somoza <gabriel@strategery.io>
  */
-class LatestCommand extends RepositoryCommand
+class LatestCommand extends AbstractRepositoryCommand
 {
     const COMMAND_NAME = 'migrations:latest';
